requirements.txt
coinspot
python-dotenv